Arreglat output del CSV d'imatge, abans printava la representació ASCII i no el valor

+++ BaseCompiled/test.cpp
@@ -24,8 +24,8 @@ void generateImageCSV(uint8_t* dades, int height, int width){
 	file.open("frameImage.csv");
 
 	for(int i = 0; i < height*width; ++i){
-		if((i % width) == (width-1)) file << dades[i] << "\n";
-		else file << dades[i] << ",";
+		if((i % width) == (width-1)) file << unsigned(dades[i]) << "\n";
+		else file << unsigned(dades[i]) << ",";
 	}
 	file.close();
 }
